requirements.txt
numpy~=1.19.5
opencv-python~=4.5.4.58
tensorflow-gpu~=2.6.0
matplotlib~=3.4.3
pygubu~=0.18
Pillow~=8.4.0